Implement recording handler
