+++ Neuro/DependencyInjection/IContextProvider.cs
@@ -1,4 +1,5 @@
 ﻿using Neuro.Pathfinding;
+using Neuro.Recording;
 using Neuro.Tasks;
 using Neuro.Visibility;
 using Neuro.Vision;
@@ -8,6 +9,7 @@ namespace Neuro.DependencyInjection;
 public interface IContextProvider
 {
     public IPathfindingHandler PathfindingHandler { get; }
+    public IRecordingHandler RecordingHandler { get; }
     public ITasksHandler TasksHandler { get; }
     public IVisibilityHandler VisibilityHandler { get; }
     public IVisionHandler VisionHandler { get; }
