+++ Neuro/NeuroPlugin.cs
@@ -4,6 +4,7 @@ using BepInEx.Unity.IL2CPP;
 using HarmonyLib;
 using Neuro.DependencyInjection;
 using Neuro.Pathfinding;
+using Neuro.Recording;
 using Neuro.Tasks;
 using Neuro.Utilities;
 using Neuro.Visibility;
@@ -20,6 +21,7 @@ public partial class NeuroPlugin : BasePlugin, IContextProvider
     public IContextProvider MainContext => this;
 
     public IPathfindingHandler PathfindingHandler { get; private set; }
+    public IRecordingHandler RecordingHandler { get; private set; }
     public ITasksHandler TasksHandler { get; private set; }
     public IVisibilityHandler VisibilityHandler { get; }
     public IVisionHandler VisionHandler { get; }
@@ -29,6 +31,7 @@ public partial class NeuroPlugin : BasePlugin, IContextProvider
         Harmony.CreateAndPatchAll(Assembly.GetExecutingAssembly(), Id);
 
         PathfindingHandler = new PathfindingHandler();
+        RecordingHandler = GameObjectUtilities.CreatePermanentSingleton<RecordingHandler>(MainContext);
         TasksHandler = GameObjectUtilities.CreatePermanentSingleton<TasksHandler>(MainContext);
     }
 }
