+++ Neuro/Recording/DataStructures/PlayerRecord.cs
@@ -1,0 +1,6 @@
+﻿using System;
+
+namespace Neuro.Recording.DataStructures;
+
+[Serializable]
+public record struct PlayerRecord(MyVector2 relativeDirection, float distance);
