+++ Neuro/Recording/IRecordingHandler.cs
@@ -1,0 +1,7 @@
+﻿using Neuro.DependencyInjection;
+
+namespace Neuro.Recording;
+
+public interface IRecordingHandler : IContextAcceptor
+{
+}
