+++ Neuro/Vision/IVisionHandler.cs
@@ -1,4 +1,5 @@
 ﻿using Neuro.DependencyInjection;
+using UnityEngine;
 
 namespace Neuro.Vision;
 
@@ -7,4 +8,6 @@ public interface IVisionHandler : IContextAcceptor
     public void ReportFindings();
 
     public void ResetAfterMeeting();
+
+    public Vector2 DirectionToNearestBody { get; }
 }
